backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "databento>=0.68.1",
    "duckdb>=1.4.3",
    "fastapi[standard]>=0.124.4",
    "nats-py>=2.12.0",
    "pandas>=2.3.3",
    "polygon-api-client>=1.16.3",
    "pyarrow>=19.0.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "s3fs>=2025.12.0",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
